fix: Make QGramIndex use less memory; fix bug; improve tests&docs (#471)
I've added some more documentation and clarifying tests to `QGramIndex`,
so that the implementation is more clear without looking at the code.

I found one small bug: q-grams of value `0` (i.e. `AAA`) wouldn't be
filtered by `max_count`. I've added a test that would fail if the
`.skip(1)` was still there.

The size of the `pos` vector was always `|text|`, but it's sufficient to
only allocate memory for the number of unfiltered q-grams. Please verify
this. I've added a `text="AAAAA"`, `pattern="AAA"` test that should show
that things are still working.